refactor and move to monai.inferers

+++ monai/inferers/utils.py
@@ -155,10 +155,10 @@ def sliding_window_inference(
             raise RuntimeError(
                 "Seems to be OOM. Please try smaller patch size or mode='constant' instead of mode='gaussian'."
             ) from e
-    importance_map = convert_data_type(importance_map, torch.Tensor, device=device, dtype=compute_dtype)[0]
+    importance_map = convert_data_type(importance_map, torch.Tensor, device=device, dtype=compute_dtype)[0]  # type: ignore
     # handle non-positive weights
     min_non_zero = max(importance_map[importance_map != 0].min().item(), 1e-3)
-    importance_map = torch.clamp(importance_map, min=min_non_zero)
+    importance_map = torch.clamp(importance_map.to(torch.float32), min=min_non_zero).to(compute_dtype)
 
     # Perform predictions
     dict_key, output_image_list, count_map_list = None, [], []
@@ -271,7 +271,7 @@ def sliding_window_inference(
         final_output = dict(zip(dict_key, output_image_list))
     else:
         final_output = tuple(output_image_list)
-    return final_output[0] if is_tensor_output else final_output
+    return final_output[0] if is_tensor_output else final_output  # type: ignore
 
 
 def _get_scan_interval(
